+++ cycpepflow/__init__.py
@@ -1,0 +1,4 @@
+from cycpepflow._version import __version__ as __version__
+from cycpepflow.models.model import BaseFlow as BaseFlow
+
+__all__ = ["BaseFlow", "__version__"]