+++ etflow/__init__.py
@@ -1,1 +1,0 @@
-from etflow.models.model import BaseFlow as BaseFlow